Fixed mixed precision issues

+++ optimizer/raven.py
@@ -116,7 +116,7 @@ class RavenAdamW(Optimizer):
 
                 if len(state) == 0:
                     state["step"] = 0
-                    state["exp_avg_cpu"] = torch.zeros_like(p, memory_format=torch.preserve_format, device='cpu', dtype=torch.bfloat16)
+                    state["exp_avg_cpu"] = torch.zeros_like(p, memory_format=torch.preserve_format, device='cpu', dtype=torch.float32)
                     state["exp_avg_sq_cpu"] = torch.zeros_like(p, memory_format=torch.preserve_format, device='cpu', dtype=torch.float32)
 
                 state["step"] += 1
